Moved firing functionality from tank to TankimingComponent
