+++ BattleTank/Source/BattleTank/Private/TankAIController.cpp
@@ -27,6 +27,6 @@ void ATankAIController::Tick(float DeltaTime)
 	auto AimingComponent = ControlledTank->FindComponentByClass<UTankAimingComponent>();
 	if (!ensure(AimingComponent)) { return; }
 	AimingComponent->AimAt(PlayerTank->GetActorLocation());
-	// ControlledTank->Fire();
+	AimingComponent->Fire();
 	
 }
