+++ BattleTank/Source/BattleTank/Public/TankAimingComponent.h
@@ -15,6 +15,7 @@ enum class EFiringState : uint8 { Reloading, Aiming, Locked };
 // Forward Declaration
 class UTankBarrel;
 class UTankTurret;
+class AProjectile;
 
 // Holds barrel's properties and elevate method
 UCLASS( ClassGroup=(Custom), meta=(BlueprintSpawnableComponent) )
@@ -24,6 +25,9 @@ class BATTLETANK_API UTankAimingComponent : public UActorComponent
 
 public:	
 	void AimAt(FVector HitLocation);
+
+	UFUNCTION(BlueprintCallable, Category="Firing")
+	void Fire();
 
 	UFUNCTION(BlueprintCallable, Category="Setup")
 	void Initialise(UTankBarrel* BarrelReference, UTankTurret* TurretReference);
@@ -36,6 +40,9 @@ private:
 	// Sets default values for this component's properties
 	UTankAimingComponent();
 
+	UPROPERTY(EditDefaultsOnly, Category="Firing")
+	bool FiringEnabled = true;
+
 	UTankBarrel* Barrel = nullptr;
 	UTankTurret* Turret = nullptr;
 
@@ -44,4 +51,11 @@ private:
 	// Sets default values for this pawn's properties
 	UPROPERTY(EditDefaultsOnly, Category="Firing")
 	float LaunchSpeed = 4000.0;
+
+	float ReloadTimeInSeconds = 3;
+	double LastFireTime = 0;
+
+	// Use TSubclassOf<> to limit what can be selected in editor. See: https://docs.unrealengine.com/latest/INT/Programming/UnrealArchitecture/TSubclassOf/
+	UPROPERTY(EditDefaultsOnly, Category="Setup")
+	TSubclassOf<AProjectile> ProjectileBlueprint;
 };
