+++ BattleTank/Source/BattleTank/Tank.cpp
@@ -1,36 +1,9 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 #include "Tank.h"
-#include "Projectile.h"
-#include "TankBarrel.h"
 
 // Sets default values
 ATank::ATank()
 {
  	// Set this pawn to call Tick() every frame.  You can turn this off to improve performance if you don't need it.
 	PrimaryActorTick.bCanEverTick = false;
-}
-
-void ATank::BeginPlay() {
-	Super::BeginPlay();
-}
-
-void ATank::Fire()
-{
-	bool isReloaded = (FPlatformTime::Seconds() - LastFireTime) > ReloadTimeInSeconds;
-
-	if (!FiringEnabled) { return; }
-	if (!ensure(Barrel)) { return; }
-
-	if (isReloaded) {
-
-		// Spawn a projectile at the socket location of the barrel
-		auto Projectile = GetWorld()->SpawnActor<AProjectile>(
-			ProjectileBlueprint,
-			Barrel->GetSocketLocation(FName("Projectile")),
-			Barrel->GetSocketRotation(FName("Projectile"))
-		);
-
-		Projectile->LaunchProjectile(LaunchSpeed);
-		LastFireTime = FPlatformTime::Seconds();
-	}
 }
